+++ public/scripts/index.js
@@ -4,7 +4,7 @@ import backgroundMusic from './backgroundMusic.js';
 import animateTeamIntroduce from './animate/animateTeamIntroduce.js';
 import showIntroduce from './showIntroduce.js'
 import clickSoundTeamMemberName from './clickSound/clickSoundTeamMemberName.js'
-import inrandomcolor from '../../test/randomFontColor.js';
+import inrandomcolor from './randomFontColor.js';
 
 showIntroduce();
 animateTeamMemberName();
